+++ lib/redmine_bots/telegram/tdlib/fetch_all_chats.rb
@@ -4,8 +4,6 @@ module RedmineBots::Telegram::Tdlib
       super
 
       @chat_list = ChatList::Main.new
-      @offset_order = 2**63 - 1
-      @offset_chat_id = 0
       @limit = 100
       @chat_futures = []
     end
@@ -17,18 +15,21 @@ module RedmineBots::Telegram::Tdlib
     private
 
     attr_reader :limit
-    attr_accessor :chat_list, :offset_order, :offset_chat_id
+    attr_accessor :chat_list
 
     def fetch
-      client.get_chats(chat_list: chat_list, offset_order: offset_order, offset_chat_id: offset_chat_id, limit: limit).then do |update|
-        chat_ids = update.chat_ids
-        next Concurrent::Promises.fulfilled_future(nil) if chat_ids.empty?
-
-        client.get_chat(chat_id: chat_ids.last).then do |chat|
-          self.offset_chat_id, self.offset_order = chat.id, chat.positions.find { |p| p.list.is_a?(ChatList::Main) }.order
+      client.load_chats(chat_list: chat_list, limit: limit).then do |update|
+        case update
+        when TD::Types::Ok
           fetch.wait!
-        end.flat
-      end.flat
+        else
+          next Concurrent::Promises.fulfilled_future(nil)
+        end
+      end.flat.rescue do |error|
+        if error.code != 404
+          raise error
+        end
+      end
     end
   end
 end